+++ source/backend/plugin/Lv2Plugin.cpp
@@ -1311,7 +1311,7 @@ public:
                 {
                     if (pData->osc.data.target != nullptr)
                     {
-                        QByteArray chunk((const char*)atom, atom->size);
+                        QByteArray chunk((const char*)atom, (int)atom->size);
                         osc_send_lv2_atom_transfer(pData->osc.data, portIndex, chunk.toBase64().constData());
                     }
                 }
@@ -1517,7 +1517,7 @@ public:
         {
             pData->param.createNew(params, true);
             fParamBuffers = new float[params];
-            FloatVectorOperations::clear(fParamBuffers, params);
+            FloatVectorOperations::clear(fParamBuffers, static_cast<int>(params));
         }
 
         if (const uint32_t count = static_cast<uint32_t>(evIns.count()))
@@ -2439,7 +2439,7 @@ public:
         {
             // disable any output sound
             for (uint32_t i=0; i < pData->audioOut.count; ++i)
-                FloatVectorOperations::clear(outBuffer[i], frames);
+                FloatVectorOperations::clear(outBuffer[i], static_cast<int>(frames));
             return;
         }
 
@@ -2551,7 +2551,7 @@ public:
             if (pData->latency > 0)
             {
                 for (uint32_t i=0; i < pData->audioIn.count; ++i)
-                    FloatVectorOperations::clear(pData->latencyBuffers[i], pData->latency);
+                    FloatVectorOperations::clear(pData->latencyBuffers[i], static_cast<int>(pData->latency));
             }
 #endif
 
@@ -3132,7 +3132,7 @@ public:
                 if (pData->latency <= frames)
                 {
                     for (uint32_t i=0; i < pData->audioIn.count; ++i)
-                        FloatVectorOperations::copy(pData->latencyBuffers[i], inBuffer[i]+(frames-pData->latency), pData->latency);
+                        FloatVectorOperations::copy(pData->latencyBuffers[i], inBuffer[i]+(frames-pData->latency), static_cast<int>(pData->latency));
                 }
                 else
                 {
@@ -3324,19 +3324,19 @@ public:
         // Set audio buffers
 
         for (uint32_t i=0; i < pData->audioIn.count; ++i)
-            FloatVectorOperations::copy(fAudioInBuffers[i], audioInBuf[i]+timeOffset, frames);
+            FloatVectorOperations::copy(fAudioInBuffers[i], audioInBuf[i]+timeOffset, static_cast<int>(frames));
 
         for (uint32_t i=0; i < pData->audioOut.count; ++i)
-            FloatVectorOperations::clear(fAudioOutBuffers[i], frames);
+            FloatVectorOperations::clear(fAudioOutBuffers[i], static_cast<int>(frames));
 
         // --------------------------------------------------------------------------------------------------------
         // Set CV buffers
 
         for (uint32_t i=0; i < fCvIn.count; ++i)
-            FloatVectorOperations::copy(fCvInBuffers[i], cvInBuf[i]+timeOffset, frames);
+            FloatVectorOperations::copy(fCvInBuffers[i], cvInBuf[i]+timeOffset, static_cast<int>(frames));
 
         for (uint32_t i=0; i < fCvOut.count; ++i)
-            FloatVectorOperations::clear(fCvOutBuffers[i], frames);
+            FloatVectorOperations::clear(fCvOutBuffers[i], static_cast<int>(frames));
 
         // --------------------------------------------------------------------------------------------------------
         // Run plugin
@@ -3404,7 +3404,7 @@ public:
                     if (isPair)
                     {
                         CARLA_ASSERT(i+1 < pData->audioOut.count);
-                        FloatVectorOperations::copy(oldBufLeft, fAudioOutBuffers[i], frames);
+                        FloatVectorOperations::copy(oldBufLeft, fAudioOutBuffers[i], static_cast<int>(frames));
                     }
 
                     float balRangeL = (pData->postProc.balanceLeft  + 1.0f)/2.0f;
